--- imagic/__init__.py
@@ -1,3 +1,0 @@
-"""Module for GreatStar quirks implementations."""
-
-IMAGIC = "iMagic by GreatStar"
