+++ zhaquirks/imagic/__init__.py
@@ -1,0 +1,3 @@
+"""Module for GreatStar quirks implementations."""
+
+IMAGIC = "iMagic by GreatStar"
